requirements.txt
python-telegram-bot==20.8
aiosqlite==0.19.0
pytz==2024.1
APScheduler==3.10.4
nest-asyncio==1.6.0